+++ frontend-archive/templates/index.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html class="w-full h-full" lang="pl">
+    <head>
+        <meta charset="utf-8">
+        <meta content="IE=edge" http-equiv="X-UA-Compatible">
+        <meta content="width=device-width,initial-scale=1.0" name="viewport">
+        <link href="https://cdn.jsdelivr.net/gh/krzysztofrewak/flat-flags-iconset@latest/style.css" rel="stylesheet">
+        <link href="https://cdnjs.cloudflare.com/ajax/libs/semantic-ui/2.4.1/components/icon.min.css" rel="stylesheet">
+        <link href="https://api.mapbox.com/mapbox-gl-js/v1.6.1/mapbox-gl.css" rel="stylesheet">
+        <link href="https://api.mapbox.com/mapbox-gl-js/plugins/mapbox-gl-geocoder/v4.4.2/mapbox-gl-geocoder.css"
+                rel="stylesheet" type="text/css">
+        <link id="favicon" rel="icon">
+        <title>Zwiedzanie browarów przez państwo Rewaków</title>
+    </head>
+    <body class="w-full h-full">
+        <noscript>
+            :(
+        </noscript>
+        <div id="app"></div>
+        <!-- built files will be auto injected -->
+    </body>
+</html>
