+++ frontend-archive/views/About.vue
@@ -1,0 +1,17 @@
+<template>
+  <div>
+    <page-header subtitle="Co tu się dzieje?" title="O nas"></page-header>
+    <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Itaque libero maiores molestiae nobis perferendis, quod
+      tempora. Amet distinctio illo in, ipsa provident quasi quibusdam rem sit temporibus vel vero voluptatibus? Lorem
+      ipsum dolor sit amet, consectetur adipisicing elit. Alias iste itaque nam sed voluptatum. Accusantium aliquid
+      aperiam, cumque, eaque in inventore iste iure modi provident, quis totam unde. Aperiam, quas.</p>
+  </div>
+</template>
+
+<script>
+import PageHeader from '../components/PageHeader'
+
+export default {
+  components: { PageHeader },
+}
+</script>
