+++ frontend/App.vue
@@ -1,43 +1,163 @@
 <template>
-  <div id="app" class="h-full flex flex-col items-stretch">
-    <navigation></navigation>
-
-    <div class="flex flex-row h-full w-full bg-gray-200 overflow-y-hidden">
-      <div class="w-0 sm:w-1/2 xl:w-1/3 bg-gray-200 overflow-x-hidden overflow-y-hidden">
-        <scroll :ops="ops">
-          <div class="p-8">
-            <router-view></router-view>
+  <div class="h-screen flex bg-gray-50 overflow-hidden">
+    <TransitionRoot as="template" :show="mobileMenuOpen">
+      <Dialog as="div" static class="fixed inset-0 flex z-40 lg:hidden" @close="mobileMenuOpen = false" :open="mobileMenuOpen">
+        <TransitionChild as="template" enter="transition-opacity ease-linear duration-300" enter-from="opacity-0" enter-to="opacity-100" leave="transition-opacity ease-linear duration-300" leave-from="opacity-100" leave-to="opacity-0">
+          <DialogOverlay class="fixed inset-0 bg-gray-600 bg-opacity-75"/>
+        </TransitionChild>
+        <TransitionChild as="template" enter="transition ease-in-out duration-300 transform" enter-from="-translate-x-full" enter-to="translate-x-0" leave="transition ease-in-out duration-300 transform" leave-from="translate-x-0" leave-to="-translate-x-full">
+          <div class="relative flex-1 flex flex-col max-w-xs w-full bg-white focus:outline-none">
+            <TransitionChild as="template" enter="ease-in-out duration-300" enter-from="opacity-0" enter-to="opacity-100" leave="ease-in-out duration-300" leave-from="opacity-100" leave-to="opacity-0">
+              <div class="absolute top-0 right-0 -mr-12 pt-4">
+                <button type="button" class="ml-1 flex items-center justify-center h-10 w-10 rounded-full focus:outline-none focus:ring-2 focus:ring-inset focus:ring-white" @click="mobileMenuOpen = false">
+                  <span class="sr-only">Close sidebar</span>
+                  <XIcon class="h-6 w-6 text-white" aria-hidden="true"/>
+                </button>
+              </div>
+            </TransitionChild>
+            <div class="pt-5 pb-4">
+              <div class="flex-shrink-0 flex items-center px-4">
+                <img class="h-8 w-auto" src="https://tailwindui.com/img/logos/workflow-mark.svg?color=indigo&shade=600" alt="Workflow"/>
+              </div>
+              <nav aria-label="Sidebar" class="mt-5">
+                <div class="px-2 space-y-1">
+                  <a v-for="item in navigation" :key="item.name" :href="item.href" class="group p-2 rounded-md flex items-center text-base font-medium text-gray-600 hover:bg-gray-50 hover:text-gray-900">
+                    <component :is="item.icon" class="mr-4 h-6 w-6 text-gray-400 group-hover:text-gray-500" aria-hidden="true"/>
+                    {{ item.name }}
+                  </a>
+                </div>
+              </nav>
+            </div>
+            <div class="flex-shrink-0 flex border-t border-gray-200 p-4">
+              <a href="#" class="flex-shrink-0 group block">
+                <div class="flex items-center">
+                  <div>
+                    <img class="inline-block h-10 w-10 rounded-full" :src="user.imageUrl" alt=""/>
+                  </div>
+                  <div class="ml-3">
+                    <p class="text-base font-medium text-gray-700 group-hover:text-gray-900">
+                      {{ user.name }}
+                    </p>
+                    <p class="text-sm font-medium text-gray-500 group-hover:text-gray-700">
+                      Account Settings
+                    </p>
+                  </div>
+                </div>
+              </a>
+            </div>
           </div>
-        </scroll>
+        </TransitionChild>
+        <div class="flex-shrink-0 w-14" aria-hidden="true">
+          <!-- Force sidebar to shrink to fit close icon -->
+        </div>
+      </Dialog>
+    </TransitionRoot>
+
+    <!-- Static sidebar for desktop -->
+    <div class="hidden lg:flex lg:flex-shrink-0">
+      <div class="flex flex-col w-20">
+        <div class="flex flex-col h-0 flex-1 overflow-y-auto bg-indigo-600">
+          <div class="flex-1 flex flex-col">
+            <div class="flex-shrink-0 bg-indigo-700 py-4 flex items-center justify-center">
+              <img class="h-8 w-auto" src="https://tailwindui.com/img/logos/workflow-mark.svg?color=white" alt="Workflow"/>
+            </div>
+            <nav aria-label="Sidebar" class="py-6 flex flex-col items-center space-y-3">
+              <a v-for="item in navigation" :key="item.name" :href="item.href" class="flex items-center p-4 rounded-lg text-indigo-200 hover:bg-indigo-700">
+                <component :is="item.icon" class="h-6 w-6" aria-hidden="true"/>
+                <span class="sr-only">{{ item.name }}</span>
+              </a>
+            </nav>
+          </div>
+          <div class="flex-shrink-0 flex pb-3 text-white flex flex-col items-center">
+            <a href="https://github.com/krzysztofrewak/browary" class="flex items-center p-4 rounded-lg text-indigo-200 hover:bg-indigo-700">
+              <CodeIcon class="h-6 w-6" aria-hidden="true"/>
+            </a>
+          </div>
+        </div>
       </div>
-      <map-panel></map-panel>
+    </div>
+
+    <div class="flex-1 min-w-0 flex flex-col overflow-hidden">
+      <!-- Mobile top navigation -->
+      <div class="lg:hidden">
+        <div class="bg-indigo-600 py-2 px-4 flex items-center justify-between sm:px-6 lg:px-8">
+          <div>
+            <img class="h-8 w-auto" src="https://tailwindui.com/img/logos/workflow-mark.svg?color=white" alt="Workflow"/>
+          </div>
+          <div>
+            <button type="button" class="-mr-3 h-12 w-12 inline-flex items-center justify-center bg-indigo-600 rounded-md text-white hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-white" @click="mobileMenuOpen = true">
+              <span class="sr-only">Open sidebar</span>
+              <MenuIcon class="h-6 w-6" aria-hidden="true"/>
+            </button>
+          </div>
+        </div>
+      </div>
+
+      <main class="flex-1 flex overflow-hidden">
+        <div class="flex-1 flex xl:overflow-hidden">
+          <!-- Primary column -->
+          <section aria-labelledby="primary-heading" class="min-w-0 flex-1 h-full flex flex-col overflow-hidden lg:order-last">
+            <h1 id="primary-heading" class="sr-only">Account</h1>
+            <!-- Your content -->
+          </section>
+
+          <!-- Secondary column (hidden on smaller screens) -->
+          <aside class="hidden lg:block lg:flex-shrink-0 lg:order-first">
+            <div class="h-full relative flex flex-col w-96 border-r border-gray-200 bg-white">
+              <!-- Your content -->
+            </div>
+          </aside>
+        </div>
+      </main>
     </div>
   </div>
 </template>
 
 <script>
-import Map from '@/components/layout/Map'
-import Navigation from '@/components/layout/Navigation'
-import Scroll from 'vuescroll'
+import { ref } from 'vue'
+import { Dialog, DialogOverlay, TransitionChild, TransitionRoot } from '@headlessui/vue'
+import {
+  BookmarkAltIcon,
+  CodeIcon,
+  FireIcon,
+  HomeIcon,
+  InboxIcon,
+  MenuIcon,
+  UserIcon,
+  XIcon
+} from '@heroicons/vue/outline'
+
+const user = {
+  name: 'Emily Selman',
+  imageUrl:
+      'https://images.unsplash.com/photo-1502685104226-ee32379fefbe?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80'
+}
+const navigation = [
+  { name: 'Home', href: '#', icon: HomeIcon },
+  { name: 'Trending', href: '#', icon: FireIcon },
+  { name: 'Bookmarks', href: '#', icon: BookmarkAltIcon },
+  { name: 'Messages', href: '#', icon: InboxIcon },
+  { name: 'Profile', href: '#', icon: UserIcon }
+]
 
 export default {
   components: {
-    'map-panel': Map,
-    Navigation,
-    Scroll,
+    Dialog,
+    DialogOverlay,
+    TransitionChild,
+    TransitionRoot,
+    MenuIcon,
+    XIcon,
+    CodeIcon
   },
-  data () {
+  setup () {
+    const mobileMenuOpen = ref(false)
+
     return {
-      ops: {
-        bar: {
-          background: '#a0aec0',
-        },
-        rail: {
-          gutterOfSide: '4px',
-          gutterOfEnds: '8px',
-        },
-      },
+      user,
+      navigation,
+      mobileMenuOpen
     }
-  },
+  }
 }
 </script>
